devuan/devuan-devs.gpg: NO GUARANTEES THIS IS LEGIT
I really have no idea how to check that the key is legit or not.
With Debian I can follow all changes since .. I use Debian,
because old distros get updated with the keys for the next one.

With Devuan I really have no idea how to verify the keys there yet.
Perhaps somebody can enlighten me how to check new keys with
the previous ones by commandline.  (With Debian this is not necessary,
as the keys are already installed in the system.)